Improve pre-commit script
Make pre-commit still work even with unstashed changes. We do this by
copying the contents of the index into a temp dir.

One bug I've noticed with this is that git commit --amend returns an
error. A work around is to do git commit -a --amend, even if there is
nothing new to add.