Move tests into src directory
Move old tests into src directory so that they work again, and use
React's test command so that we run React tests as well as the old
tests.

Tests still don't work because we can't load json5 in tests. See
https://github.com/dominique-mueller/create-react-app-typescript-json5-setup/issues/1

Also, delete pathOfficial.js, which we weren't using. It's still in Git
history.