Add support for NJ Transit buses
NJ Transit buses don't have a realtime API. They have an endpoint that
returns XML with realtime departures, but it doesn't set the CORS header
and also uses Cloudflare to prevent bot access. So to access it
programmatically, we have to run a proxy that scrapes their site using
Selenium and returns the XML.  Unfortunately this means a relatively
high barrier to entry if you want to have this service in your transit
panel config; you need some server to run the proxy on.